+++ src/core/scopes.cc
@@ -29,9 +29,9 @@
 #include "waflz/acl.h"
 #include "waflz/rules.h"
 #include "waflz/engine.h"
-#include "waflz/limit/rl_obj.h"
-#include "waflz/limit/limit.h"
-#include "waflz/limit/enforcer.h"
+#include "waflz/rl_obj.h"
+#include "waflz/limit.h"
+#include "waflz/enforcer.h"
 #include "waflz/challenge.h"
 #include "support/ndebug.h"
 #include "support/base64.h"
