+++ src/db/redis_db.cc
@@ -24,7 +24,7 @@
 //: includes
 //: ----------------------------------------------------------------------------
 #include "support/ndebug.h"
-#include "waflz/db/redis_db.h"
+#include "waflz/redis_db.h"
 #include "waflz/def.h"
 #include <stdlib.h>
 #include <string.h>
