+++ src/limit/config.cc
@@ -32,12 +32,12 @@
 #include "jspb/jspb.h"
 #include "cityhash/city.h"
 #include "waflz/def.h"
-#include "waflz/db/kv_db.h"
+#include "waflz/kv_db.h"
 #include "waflz/rqst_ctx.h"
 #include "waflz/render.h"
 #include "waflz/scopes.h"
-#include "waflz/limit/enforcer.h"
-#include "waflz/limit/config.h"
+#include "waflz/enforcer.h"
+#include "waflz/config.h"
 #include "rapidjson/document.h"
 #include "rapidjson/error/error.h"
 #include "rapidjson/error/en.h"
