+++ src/limit/configs.cc
@@ -27,9 +27,9 @@
 #include "support/file_util.h"
 #include "support/string_util.h"
 #include "support/ndebug.h"
-#include "waflz/limit/configs.h"
-#include "waflz/limit/config.h"
-#include "waflz/db/kycb_db.h"
+#include "waflz/configs.h"
+#include "waflz/config.h"
+#include "waflz/kycb_db.h"
 #include "rapidjson/document.h"
 #include "rapidjson/error/error.h"
 #include "rapidjson/error/en.h"
