+++ tests/whitebox/limit/wb_config_bc.cc
@@ -29,10 +29,10 @@
 #include "support/string_util.h"
 #include "support/ndebug.h"
 #include "waflz/def.h"
-#include "waflz/limit/config.h"
-#include "waflz/limit/configs.h"
+#include "waflz/config.h"
+#include "waflz/configs.h"
 #include "waflz/challenge.h"
-#include "waflz/db/kycb_db.h"
+#include "waflz/kycb_db.h"
 #include "waflz/rqst_ctx.h"
 #include "waflz/geoip2_mmdb.h"
 #include "limit.pb.h"
