+++ tests/whitebox/limit/wb_integration.cc
@@ -27,9 +27,9 @@
 #include "jspb/jspb.h"
 #include "support/time_util.h"
 #include "waflz/def.h"
-#include "waflz/limit/enforcer.h"
-#include "waflz/limit/config.h"
-#include "waflz/db/kycb_db.h"
+#include "waflz/enforcer.h"
+#include "waflz/config.h"
+#include "waflz/kycb_db.h"
 #include "waflz/rqst_ctx.h"
 #include "waflz/geoip2_mmdb.h"
 #include "limit.pb.h"
