+++ util/scopez_server/sx_scopes.cc
@@ -26,8 +26,8 @@
 #include "sx_scopes.h"
 #include "waflz/engine.h"
 #include "waflz/rqst_ctx.h"
-#include "waflz/db/kycb_db.h"
-#include "waflz/db/redis_db.h"
+#include "waflz/kycb_db.h"
+#include "waflz/redis_db.h"
 #include "is2/support/trace.h"
 #include "is2/support/nbq.h"
 #include "is2/support/ndebug.h"
