+++ util/waflz_server/sx_limit.cc
@@ -25,10 +25,10 @@
 //: ----------------------------------------------------------------------------
 #include "sx_limit.h"
 #include "is2/support/ndebug.h"
-#include "waflz/db/kycb_db.h"
-#include "waflz/db/redis_db.h"
-#include "waflz/limit/configs.h"
-#include "waflz/limit/config.h"
+#include "waflz/kycb_db.h"
+#include "waflz/redis_db.h"
+#include "waflz/configs.h"
+#include "waflz/config.h"
 #include "waflz/geoip2_mmdb.h"
 #include "waflz/challenge.h"
 #include "support/file_util.h"
